only load a single sheet language at a time, cache sheets/rows

+++ Lumina/Data/Language.cs
@@ -5,7 +5,7 @@ using System.ComponentModel.DataAnnotations;
 namespace Lumina.Data
 {
     [Flags]
-    public enum Language : byte
+    public enum Language : ushort
     {
         [Display( Name = "" )]
         None,
